### 2023-09-07 #5814 @kthyng
Hi @joaquinperaza and thanks for your submission. Because of the relatively small number of lines of Python code, I am going to ask the greater editorial board to consider if the package is in scope in terms of the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This process will take 1-2 weeks. Thanks.
