### 2023-09-07 #5814 @joaquinperaza
Thank you for your prompt response and valuable feedback on our manuscript submission to the Journal of Open Source Software. We appreciate the diligence of the editorial board in evaluating our work. We would like to provide some additional context regarding our CRNPy library and address the concerns raised:

CRNPy is a comprehensive library encompassing the full spectrum of routines required for the correction and analysis of raw cosmic-ray neutron data recorded by both stationary and roving devices for soil moisture monitoring. It is noteworthy that the CRNPy library incorporates, to the best of our knowledge, nearly all existing techniques and methods in this field. This collective effort is reflected in the inclusion of methods documented across 12 peer-reviewed manuscripts spanning the past 15 years since the inception of this technique. This extensive body of work underscores the completeness and robustness of our library.

Recognizing the constraints posed by the specified limit on the total lines of code in the main file, we have gone the extra mile to create three comprehensive tutorials in the form of Jupyter Notebooks that include additional Python code and are designed to facilitate the onboarding of new users, including scientists, graduate students, and lab/field technicians. These tutorials utilize real-world data and cover a range of sensor manufacturers, ensuring that users can harness the full potential of CRNPy with ease.

We want to reassure the editorial board of our commitment to aligning with the journal's requirements and guidelines. Should there be any further questions, suggestions, or specific modifications needed to accommodate our submission, we are fully prepared to collaborate closely with the editorial office to ensure that our work meets the standards set forth by the Journal of Open Source Software.

Once again, thank you for your consideration, and we look forward to any further guidance you may provide.
