### 2023-10-24 #5814 @elbeejay
@joaquinperaza I will have a look through the repository and paper later this week and get back to you with any minor comments on my end. In the meantime, are there any reviewers you might want to recommend for this work? Please consult with co-authors and check out our [list of volunteers](https://reviewers.joss.theoj.org/lookup) to make that determination.

Thanks
