### 2023-10-25 #5814 @joaquinperaza
@elbeejay Thanks for taking the time to look at our research. From the volunteer list, Dr. Shu and Joshua Larsen have demonstrated experience in soil science and hydrology.

If it would be of assistance, we can recommend researchers on the topic of cosmic-ray neutron detection. If this aligns with the review process, could you kindly let us know what information would you require? We appreciate your guidance on this.
