### 2023-10-25 #5814 @jlarsen-usgs
@elbeejay, I took a look through the repo and the paper and I'm interested in doing a review of this submission.
