### 2023-10-31 #5814 @awickert
Hi @elbeejay,

Thanks for the note. I just looked into the package, and (other than a brief memory of seeing a poster at AGU) I have no experience with cosmogenic radionuclides used as soil-moisture sensor. This unfortunately means that I have no idea whom I could recommend to you: so sorry!

With cheer and good wishes to find a referee,
Andy
