### 2023-11-01 #5814 @jeffskwang
Hello @elbeejay,
It's nice to hear from you. Like Andy, I don't have any experience whatsoever with this science, so I don't think I'd make a good reviewer. Also, I'm not aware of any of my colleagues that are working on this problem.

Did you see: [https://github.com/danpower101/crspy](https://github.com/danpower101/crspy)? Maybe someone on that project would be good?

Jeffrey
