### 2023-11-06 #5814 @danpower101
Hi @elbeejay

I've had a look through the paper/repo and I'd be happy to review it.

Thanks,
Dan
