### 2023-11-06 #6025 @editorialbot
[REVIEW]: CRNPy: An Open-Source Python Library for Cosmic Ray Neutron Probe Data Processing
**Submitting author:** @joaquinperaza (<a href="http://orcid.org/0009-0007-9337-830X">Joaquin Peraza Rud</a>)
**Repository:** https://github.com/soilwater/crnpy
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v0.6.1
