### 2023-11-06 #6025 @elbeejay
Thanks @jlarsen-usgs and @danpower101 for agreeing to review this submission to JOSS.

This is the official "review" issue, please let me know if you have any questions about the process, I believe instructions are auto-generated by the bot in the first comment including the comment you need to make to generate your reviewer checklist. It'd be great if you could complete your first passes of the software in 6 weeks, I will have the bot send us reminders in 3 weeks time.
