### 2023-12-01 soilwater/crnpy#8 @danpower101
Referencing 
This is related to the review of https://github.com/openjournals/joss-reviews/issues/6025

Add the Schrön et al., (2017) reference for the sensing depth function (it’s implemented in the code but not referenced in the paper). Double check that all the methods you've applied in the code are referenced in the paper.

Some of the references are given incorrectly, the "implementing corrections" bullet point:

Overall description (Andreasen et al., 2017)
Atmospheric pressure (Zreda et al., 2012)
Air humidity (Rosolem et al., 2013)
Incoming Neutron Flux (Zreda et al., 2012, Hawdon et al., 2014, McJannet and Desilets 2023).

I’d suggest rephrasing this to account for the specific references and how they relate to the methods described.

Schrön, M., Köhli, M., Scheiffele, L., Iwema, J., Bogena, H. R., Lv, L., Martini, E., Baroni, G., Rosolem, R., Weimar, J., Mai, J., Cuntz, M., Rebmann, C., Oswald, S. E., Dietrich, P., Schmidt, U., and Zacharias, S.: Improving calibration and validation of cosmic-ray neutron sensors in the light of spatial sensitivity, Hydrol. Earth Syst. Sci., 21, 5009–5030, https://doi.org/10.5194/hess-21-5009-2017, 2017.

Hawdon, A., McJannet, D., and Wallace, J. 2014. Calibration and correction procedures for cosmic-ray neutron soil moisture probes located across Australia. Water Resour. Res. 50: 5029–5043.

McJannet, D.L.; Desilets, D. Incoming Neutron Flux Corrections for Cosmic-Ray Soil and Snow Sensors Using the Global Neutron Monitor Network. Water Resour. Res. 2023, 59, e2022WR033889.
