### 2023-12-01 soilwater/crnpy#10 @danpower101
Statement of Need/State of the Field
This is part of the review of the https://github.com/openjournals/joss-reviews/issues/6025

There are already some python tools that process CRNS data e.g. crspy (Power et al., 2021), and corny (Schrön 2021).

The references should be included and a little more description on how this tool differs in its approach and compares to other CRNS processing tools.

Power, D., Rico-Ramirez, M. A., Desilets, S., Desilets, D., and Rosolem, R.: Cosmic-Ray neutron Sensor PYthon tool (crspy 1.2.1): an open-source tool for the processing of cosmic-ray neutron and soil moisture data, Geosci. Model Dev., 14, 7287–7307, https://doi.org/10.5194/gmd-14-7287-2021, 2021

Schrön, M.: CORNish PASDy – COsmic-Ray Neutron flavored PASDy PASDy – Processing and Analysis of Sensor Data in pYthon [data set], https://git.ufz.de/CRNS/cornish_pasdy, last access: xxx
