### 2023-12-01 #6025 @danpower101
@elbeejay I've completed my initial review.

In summary, I think it's a nice well structured set of code that can process Cosmic-Ray Neutron Sensor raw data to soil moisture data. The documentation is clear and I was easily able to run it on my own machine and work through the examples.

I've made some more specific points as separate issues in the repository that I feel should be addressed before acceptance.

If you need clarification on anything just let me know!
