### 2023-12-01 soilwater/crnpy#11 @jlarsen-usgs
JOSS Review #6025 CRNPy
This issue contains my review for the JOSS submission openjournals/joss-reviews#6025 :

You can either address this as a single issue or spawn a new issue from each bullet point

Reviews for the paper are including in the attached pdf. Of note:

- [ ] The statement of need could be better stated. Correction and filtering methods documented in many different papers is fairly common. I suggest reframing the statement of need with regard to reproducibility and robustness.
- [ ] References in the paper are missing DOIs

Software:

- [ ] Readme.md: Include a section/link to the example problems and maybe a short description of each example problem
- [ ] Testing: It would be nice to see testing of each/most of the individual methods to make sure that the code is giving you the expected output from synthetic or literature test cases. This will also make tracking down and fixing code errors in the future much easier.
- [ ] A type checker/conversion method should be developed and used with your processing methods. This would also help with long term maintainability of the code. I noticed in multiple methods, docstrings indicate that numpy, list, or pandas could be supplied as inputs, however there were pandas specific methods in the processing that would fail with a numpy array or a list.
- [ ] `total_row_counts()` method has unused input parameters. These unused input parameters should be removed from the method.
- [ ] `correction_pressure()`: update the docstrings. Docstrings list `atm_pressure` as an input parameter, however the parameter name is `pressure`.
- [ ] `correction_humidity()`: update the docstrings. `temp` is listed as an input parameter although it is not currently one.
- [ ] `get_incoming_flux()`: unused `date_format` variable should be removed
- [ ] `smooth_1d()`: Need to add the proper reference for the Savitzky-Golay filtering method. "Savitzky and Golay (1964)".
- [ ] `sensing_depth()`: remove trailing semicolons
- [ ] `find_neutron_monitor()`: verbose parameter is missing from docstrings
- [ ] `lat_lon_to_utm()`: This method could and probably should be removed. The python package `utm` does the same thing, is robust, and does not have unnecessary dependencies.

Optional but highly recommended:

- [ ]  Lint the code for readability and uniformity. I personally like using `black` for this, but there are many python linters that will do the job.
