### 2023-12-01 #6025 @jlarsen-usgs
@elbeejay

I've finished my review and have linked it from CRNPy to this issue. Overall, the software is well documented and I was able to run the examples.

I've compiled a list of review comments for the author to address. Most of these comments are relatively minor, however I do think that testing could be improved some.
