### 2023-12-04 #6025 @elbeejay
Great, thanks @danpower101 and @jlarsen-usgs for taking the time to conduct your reviews, and doubly thanks for linking those review issues to this over-arching issue thread.

@joaquinperaza if you could take a look at the review comments to make an initial assessment and get back to us here with an estimated timeframe for your revisions that'd be great. I took a look and as both reviewers stated above the software is well documented and the examples work, and so most of the comments are fairly minor. The most time-consuming are potentially the code formatting and test suite comments.

Thanks all!
