### 2023-12-21 #6025 @joaquinperaza
Thanks for your valuable feedback @danpower101 @jlarsen-usgs, I went over the review and already started working on the suggestions.

I apologize for the delay in my response. I estimate to address them in less than 2 weeks considering the required changes.

I am planning to create a new branch to address all the mentioned issues as a new version, I can reference commits for each of them in the different issues threads, please let me know if you have any preference on this.
