### 2023-12-22 a19b041
New reference flux corrections
Addressing second part of #7
