+++ src/crnpy/crnpy.py
@@ -291,7 +291,9 @@ def correction_humidity(abs_humidity, Aref):
     fw = 1 + 0.0054*(A - Aref) # Zreda et al. 2017 Eq 6.
     return fw
 
-def correction_incoming_flux(incoming_neutrons, incoming_Ref=None):
+
+
+def correction_incoming_flux(incoming_neutrons, incoming_Ref=None, fill_na=None, Rc_method=None, Rc_site=None, site_atmdepth=None, Rc_ref=None, ref_atmdepth=None):
     r"""Correction factor for incoming neutron flux.
 
     This function corrects neutron counts for incoming neutron flux using the method described in Anderson et al. (2017). The correction is performed using the following equation:
@@ -331,7 +333,28 @@ def correction_incoming_flux(incoming_neutrons, incoming_Ref=None):
         incoming_Ref = incoming_neutrons[0]
         warnings.warn('Reference incoming neutron flux not provided. Using first value of incoming neutron flux.')
     fi = incoming_neutrons / incoming_Ref
-    fi.fillna(1.0, inplace=True)  # Use a value of 1 for days without data
+
+    if Rc_method is not None:
+        if Rc_ref is None:
+            raise ValueError('Reference cutoff rigidity not provided.')
+        if Rc_site is None:
+            raise ValueError('Site cutoff rigidity not provided.')
+
+        if Rc_method == 'McJannetandDesilets2023':
+            if latitude is None or elevation is None:
+                raise ValueError('Latitude and elevation are required inputs for McJannet and Desilets (2023) method.')
+            tau = location_factor(site_atmdepth, Rc_site, ref_atmdepth, Rc_ref)
+            fi = 1 / (tau * fi + 1 - tau)
+
+        elif Rc_method == 'Hawdonetal2014':
+            Rc_corr = -0.075 * (Rc_site - Rc_ref)+ 1.0
+            fi = (fi - 1.0) * Rc_corr + 1.0
+
+        else:
+            raise ValueError('Cutoff rigidity method not found. Valid options are: McJannetandDesilets2023, Hawdonetal2014.')
+
+    if fill_na is not None:
+        fi.fillna(fill_na, inplace=True)  # Use a value of 1 for days without data
 
     return fi
 
@@ -812,6 +835,10 @@ def cutoff_rigidity(lat,lon):
         2.52 GV (Value from NMD is 2.40 GV)
 
     References:
+        Hawdon, A., McJannet, D., & Wallace, J. (2014). Calibration and correction procedures
+        for cosmic‐ray neutron soil moisture probes located across Australia. Water Resources Research,
+        50(6), 5029-5043.
+
         Smart, D. & Shea, Matthew. (2001). Geomagnetic Cutoff Rigidity Computer Program:
         Theory, Software Description and Example. NASA STI/Recon Technical Report N.
 
@@ -833,6 +860,88 @@ def cutoff_rigidity(lat,lon):
 
     return np.round(zq,2)
 
+
+def atmospheric_depth(elevation, latitude):
+    """Function to estimate the atmospheric depth for any point on Earth according to McJannet and Desilets, 2023
+
+    This function is required in the calculation of the location-dependent reference correction proposed by McJannet and Desilets, 2023.
+
+    Args:
+        elevation (float): Elevation in meters above sea level.
+        latitude (float): Geographic latitude in decimal degrees. Value in range -90 to 90
+
+    Returns:
+        (float): Atmospheric depth in g/cm2
+
+    References:
+        Atmosphere, U. S. (1976). US standard atmosphere. National Oceanic and Atmospheric Administration.
+
[…]
