### 2024-01-04 62c878d
Added the revised sample weighting procedure (Schrön et al., 2017).
Proposed in issue #9
