+++ src/crnpy/crnpy.py
@@ -689,15 +689,18 @@ def abs_humidity(relative_humidity, temp):
     return abs_h
 
 
-def nrad_weight(h,theta,distances,depth,rhob=1.4):
+def nrad_weight(h,theta,distances,depth,rhob=1.4, method=None, p=None, Hveg=None)
     """Function to compute distance weights corresponding to each soil sample.
 
     Args:
-        h (float): Air Humidity  from 0.1  to 50 in g/m^3. When h=0, the function will skip the distance weighting.
-        theta (array or pd.Series or pd.DataFrame): Soil Moisture for each sample (0.02 - 0.50 m^3/m^3)
-        distances (array or pd.Series or pd.DataFrame): Distances from the location of each sample to the origin (0.5 - 600 m)
-        depth (array or pd.Series or pd.DataFrame): Depths for each sample (m)
-        rhob (float): Bulk density in g/cm^3
+        h (np.array or pd.Series): Air Humidity  from 0.1  to 50 in g/m^3. When h=0, the function will skip the distance weighting.
+        theta (np.array or pd.Series): Soil Moisture for each sample (0.02 - 0.50 m^3/m^3)
+        distances (np.array or pd.Series): Distances from the location of each sample to the origin (0.5 - 600 m)
+        depth (np.array or pd.Series): Depths for each sample (m)
+        rhob (np.array or pd.Series): Bulk density in g/cm^3
+        p (np.array or pd.Series): Atmospheric pressure in hPa. Required for the 'Schron_2017' method.
+        Hveg (np.array or pd.Series): Vegetation height in m. Required for the 'Schron_2017' method.
+        method (str): Method to compute the distance weights. Options are 'Kohli_2015' or 'Schron_2017'.
 
     Returns:
         (array or pd.Series or pd.DataFrame): Distance weights for each sample.
@@ -706,61 +709,152 @@ def nrad_weight(h,theta,distances,depth,rhob=1.4):
         Köhli, M., Schrön, M., Zreda, M., Schmidt, U., Dietrich, P., and Zacharias, S. (2015).
         Footprint characteristics revised for field-scale soil moisture monitoring with cosmic-ray
         neutrons. Water Resour. Res. 51, 5772–5790. doi:10.1002/2015WR017169
+
+        Schrön, M., Köhli, M., Scheiffele, L., Iwema, J., Bogena, H. R., Lv, L.,
+        Martini, E., Baroni, G., Rosolem, R., Weimar, J., Mai, J., Cuntz, M., Rebmann, C.,
+        Oswald, S. E., Dietrich, P., Schmidt, U., and Zacharias, S.: Improving calibration and
+        validation of cosmic-ray neutron sensors in the light of spatial sensitivity,
+        Hydrol. Earth Syst. Sci., 21, 5009–5030, https://doi.org/10.5194/hess-21-5009-2017, 2017.
     """
 
-    # Table A1. Parameters for Fi and D86
-    p10 = 8735;       p11 = 17.1758; p12 = 11720;      p13 = 0.00978;   p14 = 7045;      p15 = 0.003632;
[…]
