### 2024-01-04 5f28836
Unversion the crnpy.egg-info
#11
