+++ src/crnpy/crnpy.py
@@ -13,7 +13,7 @@ import numpy as np
 import pandas as pd
 import requests
 import io
-
+import utm
 
 from scipy.signal import savgol_filter
 from scipy.interpolate import griddata
@@ -108,12 +108,11 @@ def fill_missing_timestamps(df, timestamp_col='timestamp', freq='H', round_times
     return df
 
 
-def total_raw_counts(counts, nan_strategy=None, timestamp_col=None):
+def total_raw_counts(counts):
     """Compute the sum of uncorrected neutron counts for all detectors.
 
     Args:
         counts (pandas.DataFrame): Dataframe containing only the columns with neutron counts.
-        nan_strategy (str): Strategy to use for NaN values. Options are 'interpolate', 'average', or None. Default is None.
 
     Returns:
         (pandas.DataFrame): Dataframe with the sum of uncorrected neutron counts for all detectors.
@@ -235,7 +234,7 @@ def correction_pressure(pressure, Pref, L):
 
 
     Args:
-        atm_pressure (list or array): Atmospheric pressure readings. Long-term average pressure is recommended.
+        pressure (list or array): Atmospheric pressure readings. Long-term average pressure is recommended.
         Pref (float): Reference atmospheric pressure.
         L (float): Atmospheric attenuation coefficient.
 
@@ -278,7 +277,6 @@ def correction_humidity(abs_humidity, Aref):
 
     Args:
         abs_humidity (list or array): Relative humidity readings.
-        temp (list or array): Temperature readings (Celsius).
         Aref (float): Reference absolute humidity (g/m^3). The day of the instrument calibration is recommended.
 
     Returns:
@@ -377,7 +375,7 @@ def get_incoming_neutron_flux(start_date, end_date, station, utc_offset=0, expan
         Documentation available:https://www.nmdb.eu/nest/help.php#howto
     """
 
-    # Example: get_incoming_flux(station='IRKT',start_date='2020-04-10 11:00:00',end_date='2020-06-18 17:00:00')
+    # Example: get_incoming_neutron_flux(station='IRKT',start_date='2020-04-10 11:00:00',end_date='2020-06-18 17:00:00')
     # Template url = 'http://nest.nmdb.eu/draw_graph.php?formchk=1&stations[]=KERG&output=ascii&tabchoice=revori&dtype=corr_for_efficiency&date_choice=bydate&start_year=2009&start_month=09&start_day=01&start_hour=00&start_min=00&end_year=2009&end_month=09&end_day=05&end_hour=23&end_min=59&yunits=0'
 
 
@@ -388,7 +386,6 @@ def get_incoming_neutron_flux(start_date, end_date, station, utc_offset=0, expan
     # Convert local time to UTC
     start_date = start_date - pd.Timedelta(hours=utc_offset)
     end_date = end_date - pd.Timedelta(hours=utc_offset)
-    date_format = '%Y-%m-%d %H:%M:%S'
     root = 'http://www.nmdb.eu/nest/draw_graph.php?'
     url_par = [ 'formchk=1',
                 'stations[]=' + station,
@@ -466,7 +463,7 @@ def get_reference_neutron_flux(station, date = pd.to_datetime("2011-05-01")):
 """
 
     # Get flux for 2011-05-01
-    df_flux = get_incoming_flux(station, date, date + pd.Timedelta(hours=24))
+    df_flux = get_incoming_neutron_flux(station, date, date + pd.Timedelta(hours=24))
     if df_flux is None:
[…]
