### 2024-01-06 15df273
check/convert types
#11
