### 2024-01-08 e1706ac
format and doc update
