+++ build/lib/crnpy/crnpy.py
@@ -5,20 +5,19 @@
  Created by Joaquin Peraza and Andres Patrignani.
 """
 
-# Import modules
-import sys
-import warnings
+import crnpy.data as data
+import io
 import numbers
 import numpy as np
 import pandas as pd
 import requests
-import io
+import sys
+import utm
+import warnings
 
-
-from scipy.signal import savgol_filter
 from scipy.interpolate import griddata
+from scipy.signal import savgol_filter
 from scipy.special import erfcinv
-import crnpy.data as data
 
 # Define python version
 python_version = (3, 7)  # tuple of (major, minor) version requirement
@@ -42,24 +41,24 @@ def remove_incomplete_intervals(df, timestamp_col, integration_time, remove_firs
     Returns:
         (pandas.DataFrame): 
     """
-    
+
     # Check format of timestamp column
     if df[timestamp_col].dtype != 'datetime64[ns]':
         raise TypeError('timestamp_col must be datetime64. Use `pd.to_datetime()` to fix this issue.')
 
     # Check if differences in timestamps are below or above the provided integration time
     idx_delta = df[timestamp_col].diff().dt.total_seconds() != integration_time
-    
+
     if remove_first:
         idx_delta[0] = True
-        
+
     # Select rows that meet the specified integration time
     df = df[~idx_delta]
     df.reset_index(drop=True, inplace=True)
 
     # Notify user about the number of rows that have been removed
     print(f"Removed a total of {sum(idx_delta)} rows.")
-    
+
     return df
 
 
@@ -94,40 +93,39 @@ def fill_missing_timestamps(df, timestamp_col='timestamp', freq='H', round_times
     for date in date_range:
         if date not in df[timestamp_col].values:
             if verbose:
-                print('Adding missing date:',date)
-            new_line = pd.DataFrame({timestamp_col:date}, index=[-1]) # By default fills columns with np.nan
-            df = pd.concat([df,new_line])
+                print('Adding missing date:', date)
+            new_line = pd.DataFrame({timestamp_col: date}, index=[-1])  # By default fills columns with np.nan
+            df = pd.concat([df, new_line])
             counter += 1
 
     df.sort_values(by=timestamp_col, inplace=True)
     df.reset_index(drop=True, inplace=True)
-    
+
     # Notify user about the number of rows that have been removed
     print(f"Added a total of {counter} missing timestamps.")
-    
+
     return df
 
 
-def total_raw_counts(counts, nan_strategy=None, timestamp_col=None):
+def total_raw_counts(counts):
     """Compute the sum of uncorrected neutron counts for all detectors.
 
     Args:
         counts (pandas.DataFrame): Dataframe containing only the columns with neutron counts.
-        nan_strategy (str): Strategy to use for NaN values. Options are 'interpolate', 'average', or None. Default is None.
 
     Returns:
         (pandas.DataFrame): Dataframe with the sum of uncorrected neutron counts for all detectors.
     """
 
     if counts.shape[0] > 1:
-        counts = counts.apply(lambda x: x.fillna(counts.mean(axis=1)),axis=0)
+        counts = counts.apply(lambda x: x.fillna(counts.mean(axis=1)), axis=0)
 
     # Compute sum of counts
     total_raw_counts = counts.sum(axis=1)
-    
+
     # Replace zeros with NaN
     total_raw_counts = total_raw_counts.replace(0, np.nan)
-    
+
     return total_raw_counts
 
 
[…]
