### 2024-01-09 b8e984d
fixed examples + expanded tests
#11
