+++ src/crnpy.egg-info/requires.txt
@@ -2,3 +2,4 @@ numpy
 pandas
 scipy
 requests
+utm
