### 2024-01-20 dc24a22
Test functions
#11 #9 #7
