### 2024-01-20 d1a2ff4
Update crnpy.py
#8
