### 2024-01-22 a70f0cf
Examples upgraded to new 0.6.0 version
