+++ build/lib/crnpy/crnpy.py
@@ -241,6 +241,8 @@ def correction_pressure(pressure, Pref, L):
         (list): fp pressure correction factor.
 
     References:
+        Zreda, M., Shuttleworth, W. J., Zeng, X., Zweck, C., Desilets, D., Franz, T., and Rosolem, R.: COSMOS: the COsmic-ray Soil Moisture Observing System, Hydrol. Earth Syst. Sci., 16, 4079–4099, https://doi.org/10.5194/hess-16-4079-2012, 2012.
+
         M. Andreasen, K.H. Jensen, D. Desilets, T.E. Franz, M. Zreda, H.R. Bogena, and M.C. Looms. 2017. Status and perspectives on the cosmic-ray neutron method for soil moisture estimation and other environmental science applications. Vadose Zone J. 16(8). doi:10.2136/vzj2017.04.0086
     """
 
@@ -282,6 +284,8 @@ def correction_humidity(abs_humidity, Aref):
         (list): fw correction factor.
 
     References:
+        Rosolem, R., W. J. Shuttleworth, M. Zreda, T. E. Franz, X. Zeng, and S. A. Kurc, 2013: The Effect of Atmospheric Water Vapor on Neutron Count in the Cosmic-Ray Soil Moisture Observing System. J. Hydrometeor., 14, 1659–1671, https://doi.org/10.1175/JHM-D-12-0120.1.
+
         M. Andreasen, K.H. Jensen, D. Desilets, T.E. Franz, M. Zreda, H.R. Bogena, and M.C. Looms. 2017. Status and perspectives on the cosmic-ray neutron method for soil moisture estimation and other environmental science applications. Vadose Zone J. 16(8). doi:10.2136/vzj2017.04.0086
     """
     A = abs_humidity
@@ -333,10 +337,6 @@ def correction_incoming_flux(incoming_neutrons, incoming_Ref=None, fill_na=None,
         M. Andreasen, K.H. Jensen, D. Desilets, T.E. Franz, M. Zreda, H.R. Bogena, and M.C. Looms. 2017. Status and perspectives on the cosmic-ray neutron method for soil moisture estimation and other environmental science applications. Vadose Zone J. 16(8). doi:10.2136/vzj2017.04.0086
 
         McJannet, D. L., & Desilets, D. (2023). Incoming neutron flux corrections for cosmic-ray soil and snow sensors using the global neutron monitor network. Water Resources Research, 59, e2022WR033889. https://doi.org/10.1029/2022WR033889
-
-
-
-
     """
     if incoming_Ref is None and not isinstance(incoming_neutrons, type(None)):
         incoming_Ref = incoming_neutrons[0]
@@ -473,7 +473,7 @@ def get_reference_neutron_flux(station, date=pd.to_datetime("2011-05-01")):
 """
 
     # Get flux for 2011-05-01
-    df_flux = get_incoming_neutron_flux(station, date, date + pd.Timedelta(hours=24))
+    df_flux = get_incoming_neutron_flux(date, date + pd.Timedelta(hours=24), station=station)
     if df_flux is None:
         warnings.warn(f"Reference neutron flux for {station} not available. Returning NaN.")
     else:
@@ -502,7 +502,7 @@ def smooth_1d(values, window=5, order=3, method='moving_median'):
         Analytical chemistry, 36(8), 1627-1639.
     """
 
-    if not isinstance(x, pd.Series) and not isinstance(x, pd.DataFrame):
+    if not isinstance(values, pd.Series) and not isinstance(x, pd.DataFrame):
         raise ValueError('Input must be a pandas Series or DataFrame')
 
     if method == 'moving_average':
[…]
