+++ paper.md
@@ -39,7 +39,7 @@ CRNPy is a Python library developed for processing and analyzing cosmic-ray neut
 
 # Statement of Need
 
-Cosmic ray neutron probes are a valuable tool for non-invasive soil moisture sensing at the hectometer scale (e.g., typical agricultural fields), filling the gap between point-level sensors and large-scale (i.e., several kilometers) remote sensors onboard orbiting satellites. However, cleaning, processing, and analyzing CRNP data involves multiple corrections and filtering steps spread across multiple peer-reviewed articles. CRNPy simplifies these steps by providing a complete, user-friendly, and well-documented library with minimal dependencies that includes real-world examples to convert raw CRNP data into soil moisture. The library is designed to be accessible to both researchers and instrument manufacturers seeking to integrate the library within their hardware. Unlike other similar libraries, CRNPy does not require any specific naming convention for the input data or the download of large external data sources or reanalysis data.
+Cosmic ray neutron probes (CRNP) are crucial for non-invasive soil moisture sensing at the hectometer scale, bridging the gap between point-level sensors and large-scale satellite remote sensors. While processing CRNP data involves multiple corrections and filtering steps described in various peer-reviewed articles, CRNPy offers a comprehensive, user-friendly library with minimal dependencies. Unlike existing tools such as `crspy` [@power2021cosmic] which is an open-source tool for processing cosmic-ray neutron and soil moisture data, CRNPy does not necessitate specific naming conventions for input data or the download of large external data sources. Also different from `corny` [@cornish_pasdy], CRNPy is not limited to a specific CRNP device and can be used with any CRNP device that records raw neutron count roving or stationary. CRNPy is designed to be versatile and modular, allowing users to customize their workflow and adapt it to their specific research needs. 
 
 # Library features
 
@@ -50,11 +50,11 @@ The CRNPy library includes common routines for creating workflows to process raw
 - Delimited-text files can be used without the need to change column names, which increases reproducibility and minimizes human error. Each function of the CRNPy library accepts either a Numpy array or a Pandas Series, enabling a more versatile, modular, and customizable workflow that adapts to instrument outputs from different manufacturers. The CRNPy examples using multiple devices provide guidelines and a starting point for users getting started with CRNP devices.
 - Outliers can be caused by hardware malfunction, vibrations, or external neutron sources (e.g., use of nearby neutron probe soil moisture meters like the 503 Hydroprobe by Instrotek, Inc.). The CRNPy library offers several methods for outliers detection using the `is_outlier()` method, which includes a simple range detection based on user-provided lower and upper boundaries, interquartile range, z-scores, and a scaled mean absolute difference [@iglewicz1993volume].
 - The ability to compute the total raw neutron counts for CRNP devices with multiple neutron detectors.
-- Implement corrections for atmospheric pressure, air humidity [@zreda2012cosmos; @andreasen2017status], and incoming neutron flux [@rosolem2013] using the `correction_pressure()`, `correction_humidity()`, and `correction_incoming_flux()` functions. This step is necessary to account for the effects of atmospheric conditions on neutron attenuation. \autoref{fig:output_stationary}a and \autoref{fig:output_stationary}b show the impact of each correction in the raw neutron count observed with a stationary CRNP.
[…]
