+++ paper.bib
@@ -18,7 +18,7 @@
       address = "Boston MA, USA",
       volume = "14",
       number = "5",
-      doi = "https://doi.org/10.1175/JHM-D-12-0120.1",
+      doi = "10.1175/JHM-D-12-0120.1",
       pages=      "1659 - 1671",
       url = "https://journals.ametsoc.org/view/journals/hydr/14/5/jhm-d-12-0120_1.xml"
 }
@@ -195,3 +195,60 @@
   year={2020},
   publisher={Frontiers Media SA}
 }
+
+@article{schron2017improving,
+  title={Improving calibration and validation of cosmic-ray neutron sensors in the light of spatial sensitivity},
+  author={Schr{\"o}n, Martin and K{\"o}hli, Markus and Scheiffele, Lena and Iwema, Joost and Bogena, Heye R and Lv, Ling and Martini, Edoardo and Baroni, Gabriele and Rosolem, Rafael and Weimar, Jannis and others},
+  journal={Hydrology and Earth System Sciences},
+  volume={21},
+  number={10},
+  pages={5009--5030},
+  year={2017},
+  publisher={Copernicus GmbH}
+  doi={10.5194/hess-21-5009-2017}
+}
+
+@article{hawdon2014calibration,
+  title={Calibration and correction procedures for cosmic-ray neutron soil moisture probes located across Australia},
+  author={Hawdon, Aaron and McJannet, David and Wallace, Jim},
+  journal={Water Resources Research},
+  volume={50},
+  number={6},
+  pages={5029--5043},
+  year={2014},
+  publisher={Wiley Online Library},
+  doi={10.1002/2013WR015138}
+}
+
+@article{mcjannet2023incoming,
+  title={Incoming Neutron Flux Corrections for Cosmic-Ray Soil and Snow Sensors Using the Global Neutron Monitor Network},
+  author={McJannet, DL and Desilets, D},
+  journal={Water Resources Research},
+  volume={59},
+  number={4},
+  pages={e2022WR033889},
+  year={2023},
+  publisher={Wiley Online Library},
+    doi={10.1029/2022WR033889
+}
+
+@article{power2021cosmic,
+  title={Cosmic-Ray neutron Sensor PYthon tool (crspy 1.2. 1): an open-source tool for the processing of cosmic-ray neutron and soil moisture data},
+  author={Power, Daniel and Rico-Ramirez, Miguel Angel and Desilets, Sharon and Desilets, Darin and Rosolem, Rafael},
+  journal={Geoscientific Model Development},
+  volume={14},
+  number={12},
+  pages={7287--7307},
+  year={2021},
+  publisher={Copernicus GmbH},
+  doi={10.5194/gmd-14-7287-2021}
+}
+
+@misc{cornish_pasdy,
+  title = {CORNish PASDy - COsmic-Ray Neutron flavored PASDy},
+  author = {Schrön, M.},
+  howpublished = {\url{https://git.ufz.de/CRNS/cornish_pasdy}},
+  year = {Accessed: 2024},
+  note = {Processing and Analysis of Sensor Data in pYthon}
+}
+
