+++ paper.bib
@@ -204,7 +204,7 @@
   number={10},
   pages={5009--5030},
   year={2017},
-  publisher={Copernicus GmbH}
+  publisher={Copernicus GmbH},
   doi={10.5194/hess-21-5009-2017}
 }
 
