+++ paper.bib
@@ -229,7 +229,7 @@
   pages={e2022WR033889},
   year={2023},
   publisher={Wiley Online Library},
-    doi={10.1029/2022WR033889
+  doi={10.1029/2022WR033889}
 }
 
 @article{power2021cosmic,
