+++ paper.md
@@ -39,7 +39,7 @@ CRNPy is a Python library developed for processing and analyzing cosmic-ray neut
 
 # Statement of Need
 
-Cosmic ray neutron probes (CRNP) are crucial for non-invasive soil moisture sensing at the hectometer scale, bridging the gap between point-level sensors and large-scale satellite remote sensors. While processing CRNP data involves multiple corrections and filtering steps described in various peer-reviewed articles, CRNPy offers a comprehensive, user-friendly library with minimal dependencies. Unlike existing tools such as `crspy` [@power2021cosmic] which is an open-source tool for processing cosmic-ray neutron and soil moisture data, CRNPy does not necessitate specific naming conventions for input data or the download of large external data sources. Also different from `corny` [@cornish_pasdy], CRNPy is not limited to a specific CRNP device and can be used with any CRNP device that records raw neutron count roving or stationary. CRNPy is designed to be versatile and modular, allowing users to customize their workflow and adapt it to their specific research needs. 
+Cosmic ray neutron probes (CRNP) are crucial for non-invasive soil moisture sensing at the hectometer scale, bridging the gap between point-level sensors and large-scale satellite remote sensors. While processing CRNP data involves multiple corrections and filtering steps described in various peer-reviewed articles, CRNPy offers a comprehensive, user-friendly library with minimal dependencies. Unlike existing tools such as `crspy` [@power2021cosmic] which is an open-source tool for processing cosmic-ray neutron and soil moisture data, CRNPy does not necessitate specific naming conventions for input data or the download of large external data sources. Differentiating from existing tools like `corny` [@cornish_pasdy], which target CRNP data processing workflows, CRNPy introduces versatility and modularity. It not only reaches diverse research needs but also facilitates seamless integration into deployed cloud solutions or CRNP coupled with Systems-on-Chip (SoC), offering a broader range of applicability and customization.
 
 # Library features
 
