+++ paper.bib
@@ -6,7 +6,8 @@
   number={11},
   pages={4079--4099},
   year={2012},
-  publisher={Copernicus GmbH}
+  publisher={Copernicus GmbH},
+  doi={10.5194/hess-16-4079-2012}
 }
 
 @article {rosolem2013,
@@ -31,7 +32,8 @@
   number={8},
   pages={1--11},
   year={2017},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.2136/vzj2017.04.0086}
 }
 
 @INPROCEEDINGS{2009NMDB,
@@ -61,7 +63,8 @@
   volume={46},
   number={11},
   year={2010},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1029/2009WR008726}
 }
 
 @article{baatz2015empirical,
@@ -72,7 +75,8 @@
   number={4},
   pages={2030--2046},
   year={2015},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/2014WR016443}
 }
 
 @article{wahbi2018situ,
@@ -81,7 +85,8 @@
   journal={Cosmic Ray Neutron Sensing: Estimation of Agricultural Crop Biomass Water Equivalent},
   pages={5--9},
   year={2018},
-  publisher={Springer International Publishing}
+  publisher={Springer International Publishing},
+  doi={10.1007/978-3-319-69539-6_2}
 }
 
 @article{schron2018cosmic,
@@ -92,7 +97,8 @@
   number={9},
   pages={6441--6459},
   year={2018},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1029/2017WR021719}
 }
 
 @article{smart2001geomagnetic,
@@ -119,6 +125,7 @@
   pages={2199--2211},
   year={2018},
   publisher={Wiley Online Library}
+  doi={10.1002/2017WR021692}
 }
 
 @article{zreda2008measuring,
@@ -128,7 +135,8 @@
   volume={35},
   number={21},
   year={2008},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1029/2008GL035655}
 }
 
 @article{franz2012measurement,
@@ -138,7 +146,8 @@
   volume={48},
   number={8},
   year={2012},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1029/2012WR011871}
 }
 
 @article{kohli2015footprint,
@@ -149,7 +158,8 @@
   number={7},
   pages={5772--5790},
   year={2015},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/2015WR017169}
 }
 
 
@@ -160,7 +170,8 @@
   volume={2},
   pages={9},
   year={2020},
-  publisher={Frontiers Media SA}
+  publisher={Frontiers Media SA},
+  doi={10.3389/frwa.2020.00009}
 }
 
 
@@ -172,7 +183,8 @@
   number={6},
   pages={1323--1337},
   year={2008},
-  publisher={Copernicus Publications G{\"o}ttingen, Germany}
+  publisher={Copernicus Publications G{\"o}ttingen, Germany},
+  doi={10.5194/hess-12-1323-2008}
 }
 
 @article{rossini2021predicting,
@@ -183,7 +195,8 @@
   number={6},
   pages={1894--1902},
   year={2021},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/saj2.20319}
 }
 
 @article{jakobi2020error,
[…]
