+++ paper.md
@@ -35,28 +35,30 @@ editor_options:
 
 # Summary
 
-CRNPy is a Python library developed for processing and analyzing cosmic-ray neutron counts recorded by stationary and roving moderated neutron probes (CRNP) for soil moisture sensing. The CRNPy library includes common routines for implementing atmospheric corrections, biomass corrections, road corrections, one- and two-dimensional filtering, horizontal and vertical footprint analysis, uncertainty estimation, depth extrapolation operators, device calibration, and conversion of neutron counts into volumetric soil water content. The library is designed to be user-friendly, versatile, and instrument agnostic to enable users the integration of the CRNPy library in a wide range of agricultural and hydrological applications.
+CRNPy is a Python library that facilitates the processing, analysis, and conversion of raw cosmic-ray neutron counts obtained with stationary and roving cosmic-ray neutron probes (CRNP) into volumetric soil water content. The CRNPy library includes routines for atmospheric, biomass, and road corrections, along with one-dimensional and two-dimensional filtering. The library extends its utility by offering horizontal and vertical footprint determination, uncertainty estimation, depth extrapolation operators, and routines to assist users with field calibration. The design of the CRNPy library prioritizes reproducibility, ease of use, and compatibility across instruments, facilitating its adoption by manufacturers, end users, and researchers working to integrate non-invasive soil moisture sensing in agricultural and hydrological applications. 
 
 # Statement of Need
 
-Cosmic ray neutron probes (CRNP) are crucial for non-invasive soil moisture sensing at the hectometer scale, bridging the gap between point-level sensors and large-scale satellite remote sensors. While processing CRNP data involves multiple corrections and filtering steps described in various peer-reviewed articles, CRNPy offers a comprehensive, user-friendly library with minimal dependencies. Unlike existing tools such as `crspy` [@power2021cosmic] which is an open-source tool for processing cosmic-ray neutron and soil moisture data, CRNPy does not necessitate specific naming conventions for input data or the download of large external data sources. Differentiating from existing tools like `corny` [@cornish_pasdy], which target CRNP data processing workflows, CRNPy introduces versatility and modularity. It not only reaches diverse research needs but also facilitates seamless integration into deployed cloud solutions or CRNP coupled with Systems-on-Chip (SoC), offering a broader range of applicability and customization.
[…]
