+++ paper.bib
@@ -124,7 +124,7 @@
   number={3},
   pages={2199--2211},
   year={2018},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
   doi={10.1002/2017WR021692}
 }
 
