+++ paper.bib
@@ -89,17 +89,21 @@
   doi={10.1007/978-3-319-69539-6_2}
 }
 
-@article{schron2018cosmic,
-  title={Cosmic-ray neutron rover surveys of field soil moisture and the influence of roads},
-  author={Schr{\"o}n, Martin and Rosolem, Rafael and K{\"o}hli, Markus and Piussi, L and Schr{\"o}ter, I and Iwema, J and K{\"o}gler, S and Oswald, Sascha Eric and Wollschl{\"a}ger, U and Samaniego, Luis and others},
-  journal={Water Resources Research},
-  volume={54},
-  number={9},
-  pages={6441--6459},
-  year={2018},
-  publisher={Wiley Online Library},
-  doi={10.1029/2017WR021719}
+@article{schron2018roads
+author = {Schrön, M. and Rosolem, R. and Köhli, M. and Piussi, L. and Schröter, I. and Iwema, J. and Kögler, S. and Oswald, S. E. and Wollschläger, U. and Samaniego, L. and Dietrich, P. and Zacharias, S.},
+title = {Cosmic-ray Neutron Rover Surveys of Field Soil Moisture and the Influence of Roads},
+journal = {Water Resources Research},
+volume = {54},
+number = {9},
+pages = {6441-6459},
+keywords = {road effect, field-scale, soil moisture, cosmic ray neutrons, mobile survey, COSMOS rover},
+doi = {10.1029/2017WR021719},
+url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2017WR021719},
+eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2017WR021719},
+abstract = {Abstract Measurements of root-zone soil moisture across spatial scales of tens to thousands of meters have been a challenge for many decades. The mobile application of Cosmic Ray Neutron Sensing (CRNS) is a promising approach to measure field soil moisture noninvasively by surveying large regions with a ground-based vehicle. Recently, concerns have been raised about a potentially biasing influence of local structures and roads. We employed neutron transport simulations and dedicated experiments to quantify the influence of different road types on the CRNS measurement. We found that roads introduce a substantial bias in the CRNS estimation of field soil moisture compared to off-road scenarios. However, this effect becomes insignificant at distances beyond a few meters from the road. Neutron measurements on the road could overestimate the field value by up to 40 \% depending on road material, width, and the surrounding field water content. The bias could be largely removed with an analytical correction function that accounts for these parameters. Additionally, an empirical approach is proposed that can be used without prior knowledge of field soil moisture. Tests at different study sites demonstrated good agreement between road-effect corrected measurements and field soil moisture observations. However, if knowledge about the road characteristics is missing, measurements on the road could substantially reduce the accuracy of this method. Our results constitute a practical advancement of the mobile CRNS methodology, which is important for providing unbiased estimates of field-scale soil moisture to support applications in hydrology, remote sensing, and agriculture.},
+year = {2018}
 }
+
 
 @article{smart2001geomagnetic,
   title={Geomagnetic cutoff rigidity computer program: Theory, software description and example},
@@ -210,16 +214,17 @@
   doi={10.3389/frwa.2020.00010}
 }
 
+
 @article{schron2017improving,
-  title={Improving calibration and validation of cosmic-ray neutron sensors in the light of spatial sensitivity},
[…]
