### 2024-02-03 #6025 @elbeejay
@joaquinperaza - I've noticed that you've been making changes and tying your commits to reviewer issues, great! Can you let us know where those stand; are you done revising your manuscript and making changes to the software?

I think it'd be most clear if you replied with a comment in the various issues that are open to state how you've addressed it. This might help our reviewers more easily track whether the change has been made.

Thanks!
