### 2024-02-04 soilwater/crnpy#7 @joaquinperaza
@danpower101 Thanks for your valuable feedback! I implemented the suggested changes. I slightly changed the approach in the reference date comment. Our solution in this iteration was to include the method `get_reference_neutron_flux(station, date = pd.to_datetime("2011-05-01"))`, allowing by default pulling the value for the station on 1 May 2011, but if needed, we ensure the library is still flexible, allowing to specify a different date, e.g. due to having a different calibration approach, NM availability, etc.
