### 2024-02-04 soilwater/crnpy#9 @joaquinperaza
@danpower101 Please feel free to check if the last changes in the software addressed your suggestions.
