### 2024-02-04 soilwater/crnpy#11 @joaquinperaza
@jlarsen-usgs Thanks for your valuable feedback! I updated the library software in the previous commits to address your suggestions.
