### 2024-02-17 #6025 @joaquinperaza
@elbeejay, I just submitted the last revision of the manuscript to https://github.com/soilwater/crnpy/tree/JOSS. Addressing the comments about the text from [#8](https://github.com/soilwater/crnpy/issues/8), [#10](https://github.com/soilwater/crnpy/issues/10),  [#11](https://github.com/soilwater/crnpy/issues/11) . The software changes were all implemented in the branch [`0.6.0`](https://github.com/soilwater/crnpy/tree/0.6.0)

Please let me know if anything else is needed from our side, and thanks for your support!
