### 2024-02-19 #6025 @elbeejay
Great, thanks @joaquinperaza - I'd ask @jlarsen-usgs and @danpower101, if you both could review the changes relevant to your open issues and confirm that your comments have been addressed that would be great. Once the issues are closed (and presumably the branch is merged) we can discuss the next steps @joaquinperaza.

I'll have the bot render the new PDF below for our review.
