### 2024-02-28 #6025 @elbeejay
@joaquinperaza to make things easier on @jlarsen-usgs and @danpower101 would you mind summarizing your revisions in response to their comments in the appropriate issues? Thanks - and @jlarsen-usgs and @danpower101 when y'all get a chance if you could check and potentially close the issues you'd opened that'd be great
