### 2024-03-03 soilwater/crnpy#8 @joaquinperaza
References were added into the code, documentation and paper as suggested
