### 2024-03-03 soilwater/crnpy#10 @joaquinperaza
The statement of need was updated to include previous developments,

Power, D., Rico-Ramirez, M. A., Desilets, S., Desilets, D., and Rosolem, R.: Cosmic-Ray neutron Sensor PYthon tool (crspy 1.2.1): an open-source tool for the processing of cosmic-ray neutron and soil moisture data, Geosci. Model Dev., 14, 7287–7307, https://doi.org/10.5194/gmd-14-7287-2021, 2021

Schrön, M.: CORNish PASDy – COsmic-Ray Neutron flavored PASDy PASDy – Processing and Analysis of Sensor Data in pYthon [data set], https://git.ufz.de/CRNS/cornish_pasdy, last access: xxx

and how this library differs from them in the following comit to the `JOSS` branch https://github.com/soilwater/crnpy/commit/e257e79f1cecab6b75b23ad7e76988c38b7ae9bf
