### 2024-03-03 soilwater/crnpy#11 @joaquinperaza
A series of commits were made to address the feedback comprehensively. The crnpy.egg-info folder was unversioned from the repository, aligning with best practices for Python package repositories. The README.md was updated to include links and descriptions of example problems, enhancing the documentation for users. Extensive testing was added, covering individual methods to ensure they produce expected outputs, which aids in future debugging and code maintenance. A type checker/conversion method was implemented to ensure robust handling of different input types, addressing the issue of method failures due to incompatible input types.

Unused input parameters in methods like total_row_counts() were removed, streamlining the code. Docstrings for methods such as correction_pressure() and correction_humidity() were updated for accuracy. The unused date_format variable in the get_incoming_flux() method was removed, and the proper reference for the Savitzky-Golay filtering method was added to smooth_1d(), adhering to academic standards. Trailing semicolons were removed from sensing_depth(), and the missing verbose parameter was added to the docstrings of find_neutron_monitor(). The lat_lon_to_utm() method was refined, considering the availability of the robust utm Python package, which serves a similar purpose.

The code was formatted to increase readability, and the statement of need was updated considering changes also from issue #10, References and DOI's were fixed together with comments from #8
