### 2024-03-03 4636770
Code formatting of new methods
#11
