### 2024-03-05 soilwater/crnpy#11 @elbeejay
@jlarsen-usgs if you could take another look and weigh in when you get a chance that'd be great, thanks!
