### 2024-03-18 soilwater/crnpy#10 @danpower101
Mostly addressed, although one minor comment. The point about corny gives the impression it's only a GUI based software, but it does have a command line interface too. A minor adjustment here to reflect this wouldn't impact the rest of the text.

I would suggest corny is designed as a more integrated software (I define "integrated" here to mean that it runs all the steps from a single function call), whereas crnpy is more function based/modular (e.g., it can be added to jupyter notebooks or added to processing pipelines).
