### 2024-03-24 #6025 @elbeejay
Thanks @danpower101 for your follow-up comments in both https://github.com/soilwater/crnpy/issues/9 and https://github.com/soilwater/crnpy/issues/10. @jlarsen-usgs when you get a chance if you could revisit https://github.com/soilwater/crnpy/issues/11 that'd be much appreciated.
