### 2024-04-07 cfb5684
Fixed rscaled with default Hveg to at least account for pressure without biomass data
Addressing 2nd sets of comments in #9
