### 2024-04-08 154f85d
Fixed Weighting according to Schrön et al. (2017)
Addressing the comments in #9, the method was updated to use the iterative routine.

Examples and tests, were also updated.
