+++ build/lib/crnpy/crnpy.py
@@ -670,7 +670,7 @@ def sensing_depth(vwc, pressure, p_ref, bulk_density, Wlat, dist=None, method='S
 
             # Compute soil depth that accounts for 86% of the neutron flux
             D86 = 1 / bulk_density * (8.321 + 0.14249 * (0.96655 + np.exp(-0.01 * r_start)) * (20 + (Wlat + vwc)) / (
-                        0.0429 + (Wlat + vwc)))
+                    0.0429 + (Wlat + vwc)))
             results.append(D86)
 
     elif method == 'Franz_2012':
@@ -706,7 +706,7 @@ def abs_humidity(relative_humidity, temp):
     return abs_h
 
 
-def nrad_weight(h, theta, distances, depth, rhob=1.4, method="Kohli_2015", p=None, Hveg=None):
+def nrad_weight(h, theta, distances, depth, rhob=1.4, method="Kohli_2015", p=None, Hveg=0, tol=0.01):
     """Function to compute distance weights corresponding to each soil sample.
 
     Args:
@@ -718,9 +718,11 @@ def nrad_weight(h, theta, distances, depth, rhob=1.4, method="Kohli_2015", p=Non
         p (np.array or pd.Series): Atmospheric pressure in hPa. Required for the 'Schron_2017' method.
         Hveg (np.array or pd.Series): Vegetation height in m. Required for the 'Schron_2017' method.
         method (str): Method to compute the distance weights. Options are 'Kohli_2015' or 'Schron_2017'.
+        tol (float): Tolerance for the iterative solution. Default is 0.01. Required for the 'Schron_2017' method.
 
     Returns:
-        (array or pd.Series or pd.DataFrame): Distance weights for each sample.
+        theta_new (np.array or pd.Series): Weighted soil moisture values.
+        weights (np.array or pd.Series): Distance weights for each sample. For the 'Schron_2017' method, the weights are computed for each distance.
 
     References:
         Köhli, M., Schrön, M., Zreda, M., Schmidt, U., Dietrich, P., and Zacharias, S. (2015).
@@ -826,7 +828,10 @@ def nrad_weight(h, theta, distances, depth, rhob=1.4, method="Kohli_2015", p=Non
 
         # Combined and normalized weights
         weights = Wd * W / np.nansum(Wd * W)
-        return weights
+
+        theta_new = np.sum(theta * weights)
+
+        return theta_new, weights
     elif method == 'Schron_2017':
         # Horizontal distance weights According to Eq. 6 and Table A1 in Schrön et al. (2017)
         # Method for calculating the horizontal distance weights from 0 to 1m
@@ -929,45 +934,70 @@ def nrad_weight(h, theta, distances, depth, rhob=1.4, method="Kohli_2015", p=Non
 
             return B0 * np.exp(-B1 * r) + B2 * np.exp(-B3 * r)
 
-        def rscaled(r, p, Hveg, y):
+        # Wrapper method for calculating the horizontal distance weights
+        def Wr(r, x, y):
+            if r <= 1:
+                return WrX(r, x, y)
+            elif r <= 50:
+                return WrA(r, x, y)
+            elif r <= 600:
+                return WrB(r, x, y)
+            else:
+                raise ValueError("r must be between 1 and 600m when using 'Schron_2017' method")
+
+        def rscaled(r, p, y, Hveg = 0):
             Fp = 0.4922 / (0.86 - np.exp(-p / 1013.25))
             Fveg = 1 - 0.17 * (1 - np.exp(-0.41 * Hveg)) * (1 + np.exp(-9.25 * y))
             return r / Fp / Fveg
 
         # Rename variables to be consistent with the revised paper
         r = distances
-        x = h
-        y = theta
[…]
