### 2024-04-08 f27f83c
Addressing commments in #10
- Change the definition about `corny`
