+++ paper.md
@@ -39,7 +39,7 @@ CRNPy is a Python library that facilitates the processing, analysis, and convers
 
 # Statement of Need
 
-Cosmic ray neutron probes (CRNP) are non-invasive soil moisture sensors that fill the niche between point-level and satellite sensors. However, the conversion of raw CRNP data into soil moisture requires multiple corrections and filtering steps that are described across various peer-reviewed articles. To circumvent this limitation and enhance reproducibility, the CRNPy library offers a simple, instrument-agnostic, and integrated solution with minimal dependencies. Compared to the existing `crspy`[@power2021cosmic] library, CRNPy avoids stringent data naming conventions and external data requirements. A flexible data naming convention enables a seamless integration with output files from different instrument manufacturers and the lack of external data requirements makes the library more compact (only ~65 KB)  and straight forward to install. Compared to the `corny` [@cornish_pasdy] toolbox's GUI-based workflow, CRNPy's modular design based on Python functions promotes integration and reproducibility within data analysis pipelines and interactive development environments like Jupyter Lab notebooks. In addition, its straightforward installation using the Python Package Index, the minimal dependencies—most included with the Anaconda open-source ecosystem—and the comprehensive datasets with included examples in the form of Jupyter notebooks, provide an accessible start for CRNP data processing. The CRNPy library emphasizes easy maintenance and community-driven improvements since users can expand its capabilities by adding regular Python functions to the core module. The compact size and simple structure of the CRNPy library can also enable future integration into cloud-based services, IoT sensors, and system-on-chip technologies, broadening its use and customization potential.
+Cosmic ray neutron probes (CRNP) are non-invasive soil moisture sensors that fill the niche between point-level and satellite sensors. However, the conversion of raw CRNP data into soil moisture requires multiple corrections and filtering steps that are described across various peer-reviewed articles. To circumvent this limitation and enhance reproducibility, the CRNPy library offers a simple, instrument-agnostic, and integrated solution with minimal dependencies. Compared to the existing `crspy`[@power2021cosmic] library, CRNPy avoids stringent data naming conventions and external data requirements. A flexible data naming convention enables a seamless integration with output files from different instrument manufacturers and the lack of external data requirements makes the library more compact (only ~65 KB)  and straight forward to install. Compared to the `corny` [@cornish_pasdy], a more integrated software implementing an end-to-end solution, CRNPy's modular design based on Python functions promotes integration and reproducibility within data analysis pipelines and interactive development environments like Jupyter Lab notebooks. In addition, its straightforward installation using the Python Package Index, the minimal dependencies—most included with the Anaconda open-source ecosystem—and the comprehensive datasets with included examples in the form of Jupyter notebooks, provide an accessible start for CRNP data processing. The CRNPy library emphasizes easy maintenance and community-driven improvements since users can expand its capabilities by adding regular Python functions to the core module. The compact size and simple structure of the CRNPy library can also enable future integration into cloud-based services, IoT sensors, and system-on-chip technologies, broadening its use and customization potential.
 
 # Library features
 
