### 2024-04-09 soilwater/crnpy#10 @mschroen
Just a minor comment on Corny: it can definitely be used as an integrated end-to-end processor, raw data in, plots out. And admitteldy, this is the major application method for users so far. The GUI is just an add-on that is used by 50% of the users. Others simply use the console, as it runs the same procedures but much faster.

However, Corny was made many years ago with the same purpose in mind, to be modular in a sense that every single function can be used outside of Corny, in Jupyter notebooks , etc.. People are using this feature all the time in their Notebooks, e.g.:
```python
from corny.grains.Koehli2021fiw import N2SM_Desilets
N2SM_Desilets(N=700, N0=1000)
# 0.131
```
You may also want to look at Corny's `example/` folder which includes python scripts which are essentially interactive notebooks.

Naturally, Corny is not perfect at all, and just as crnpy, heavily relies on a user community that actively develops it. I'd appreciate any effort made to implement modularizaton better. But I wanted to make clear that Corny is *neither* just a GUI *nor* a black-box! It rather is a community-driven collection of useful and extensible python functions to do CRNS processing, in the first place! It'd be good if that could be communicated adequately in publications :-)
