### 2024-04-11 #6025 @elbeejay
Thanks @andres-patrignani and @joaquinperaza for making revisions.

Checking in here with @danpower101 and @jlarsen-usgs.

@danpower101 it looks like some additional changes have been made in https://github.com/soilwater/crnpy/issues/9

@jlarsen-usgs if you could revisit https://github.com/soilwater/crnpy/issues/11 when you get a chance that'd be great.

Thank you both for reviewing and continuing to engage with the submitting authors throughout this process!
