### 2024-04-19 soilwater/crnpy#9 @danpower101
I was able to work with the .ipynb notebook for the calibration (which is working).

I noticed however that the profile is being created using the distance from the sensor, which it shouldn't. Doing it this way will mean, for example, that a profile taken at North 25m will be combined with a profile taken at South 25m and treated as a single profile. Better to use individual profile assignments to make sure they're treated separately.
