### 2024-04-19 soilwater/crnpy#10 @danpower101
An update to the text seems appropriate here considering the comment from @mschroen. As he has shown, there are already instances where corny has also been used in Jupyter notebook pipelines and so the text in it's current form could give a false impression on corny.
