### 2024-04-19 19d1549
Updated sentence about Corny library
