### 2024-04-19 soilwater/crnpy#10 @andres-patrignani
@danpower101 I updated the sentence about the Corny library in the JOSS paper using similar words to those of @mschroen. I noticed that the Corny library is a bit harder to find online whereas CRNPy usually appears in the first page of results in Google and Bing search engines. All three libraries (crspy, Corny, and CRNPy) are similar, the two main aspects that we highlighted for CRNPy are the modularity and the online documentation. Please, let us know if you have a different opinion.
