### 2024-04-19 472a25c
Minor fixes in code style
