+++ build/lib/crnpy/crnpy.py
@@ -1,1613 +1,0 @@
-# crnpy/crnpy.py
-"""
-`crnpy` is a Python package for processing cosmic ray neutron data.
-
- Created by Joaquin Peraza and Andres Patrignani.
-"""
-
-import crnpy.data as data
-import io
-import numbers
-import numpy as np
-import pandas as pd
-import requests
-import sys
-import utm
-import warnings
-
-from scipy.interpolate import griddata
-from scipy.signal import savgol_filter
-from scipy.special import erfcinv
-
-# Define python version
-python_version = (3, 7)  # tuple of (major, minor) version requirement
-python_version_str = str(python_version[0]) + "." + str(python_version[1])
-
-# produce an error message if the python version is less than required
-if sys.version_info < python_version:
-    msg = "Module only runs on python version >= %s" % python_version_str
-    raise Exception(msg)
-
-
-def remove_incomplete_intervals(df, timestamp_col, integration_time, remove_first=False):
-    """Function that removes rows with incomplete integration intervals.
-    
-    Args:
-        df (pandas.DataFrame): Pandas Dataframe with data from stationary or roving CRNP devices.
-        timestamp_col (str): Name of the column with timestamps in datetime format.
-        integration_time (int): Duration of the neutron counting interval in seconds. Typical values are 60 seconds and 3600 seconds.
-        remove_first (bool, optional): Remove first row. Default is False.
-        
-    Returns:
-        (pandas.DataFrame): 
-    """
-
-    # Check format of timestamp column
-    if df[timestamp_col].dtype != 'datetime64[ns]':
-        raise TypeError('timestamp_col must be datetime64. Use `pd.to_datetime()` to fix this issue.')
-
-    # Check if differences in timestamps are below or above the provided integration time
-    idx_delta = df[timestamp_col].diff().dt.total_seconds() != integration_time
-
-    if remove_first:
-        idx_delta[0] = True
-
-    # Select rows that meet the specified integration time
-    df = df[~idx_delta]
-    df.reset_index(drop=True, inplace=True)
-
-    # Notify user about the number of rows that have been removed
-    print(f"Removed a total of {sum(idx_delta)} rows.")
-
-    return df
-
-
-def fill_missing_timestamps(df, timestamp_col='timestamp', freq='H', round_timestamp=True, verbose=False):
-    """Helper function to fill rows with missing timestamps in datetime record. Rows are filled with NaN values.
-
-     Args:
-         df (pandas.DataFrame): Pandas DataFrame.
-         timestamp_col (str, optional): Column with the timestamp. Must be in datetime format. Default column name is 'timestamp'.
-         freq (str, optional): Timestamp frequency. 'H' for hourly, 'M' for minute, or None. Can also use '3H' for a 3 hour frequency. Default is 'H'.
-         round_timestamp (bool, optional): Whether to round timestamps to the nearest frequency. Default is True.
-         verbose (bool, optional): Prints the missing timestamps added to the DatFrame.
-
-     Returns:
-         (pandas.DataFrame): DataFrame with filled missing timestamps.
-
-     """
-
-    # Check format of timestamp column
-    if df[timestamp_col].dtype != 'datetime64[ns]':
-        raise TypeError('timestamp_col must be datetime64. Use `pd.to_datetime()` to fix this issue.')
-
-    # Round timestamps to nearest frequency. This steps must preced the filling of rows.
-    if round_timestamp:
[…]
