### 2024-04-21 soilwater/crnpy#11 @elbeejay
@jlarsen-usgs - let us know when you'll have a chance to review these changes
