### 2024-04-21 soilwater/crnpy#10 @elbeejay
@joaquinperaza and @andres-patrignani a few comments regarding the statement of need portion of the paper:

- There appears to be a space missing in between `crspy` and its citation: "`crspy`[@power2021cosmic]" should be "`crspy` [@power2021cosmic]"
- You should capitalize "Python" as you do elsewhere in the section and paper; the revised sentence includes an instance of "python" instead
- I would encourage alternative phrasing from the two new points introduced. Looking at the `corny` package and @mschroen's comment above, `corny` is also modular and extensible. From the JOSS perspective, the statement of need only needs to make clear the purpose of the software and its context in the field (refer back to the [JOSS docs](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)). We do not require a point-by-point comparison or detailed critique of alternative software packages.
