### 2024-04-21 soilwater/crnpy#9 @joaquinperaza
@danpower101, I fixed the method to require a unique profile indetifier, in the example I created a new column `ID` to compensate for the lack of that information in the sample dataset, also the unit tests were updated to account for this additional input.
