### 2024-04-21 0692b6b
Fixed bug in `smooth_1d()`
