### 2024-04-21 soilwater/crnpy#10 @andres-patrignani
@elbeejay I updated the Statement of Need. I decided to remove the comparison with the other libraries and focus on the contributions of CRNPy. Eventually users can decide what library they want to use. If the editor or reviewers prefer a different approach, please send @joaquinperaza and I a tentative sentence and we will incorporate it into the paper.

I believe we have addressed and implemented all the suggested changes made by the editors and the reviewers.
