+++ paper.md
@@ -39,7 +39,7 @@ CRNPy is a Python library that facilitates the processing, analysis, and convers
 
 # Statement of Need
 
-Cosmic ray neutron probes (CRNP) are non-invasive soil moisture sensors that fill the niche between point-level and satellite sensors. However, the conversion of raw CRNP data into soil moisture requires multiple corrections and filtering steps that are described across various peer-reviewed articles. To circumvent this limitation and enhance reproducibility, the CRNPy library offers a simple, modular, and instrument-agnostic solution that promotes integration and reproducibility within data analysis pipelines. In addition, CRNPy has a straightforward installation using the Python Package Index and minimal dependencies that are mostly included with the Anaconda open-source ecosystem. CRNPy's web documentation includes actual datasets and tutorials in the form of Jupyter notebooks that provide new users with an easily accessible entry point for CRNP data processing. The CRNPy library emphasizes easy maintenance and community-driven improvements since users can expand its capabilities by adding regular Python functions to the core module. The compact size and simple structure of the CRNPy library can also enable future integration into cloud-based services, IoT sensors, and system-on-chip technologies, broadening its use and customization potential.
+Cosmic ray neutron probes (CRNP) are non-invasive soil moisture sensors that fill the niche between point-level and satellite sensors. However, the conversion of raw CRNP data into soil moisture requires multiple corrections and filtering steps that are described across various peer-reviewed articles. To circumvent this limitation and enhance reproducibility, the CRNPy library offers a simple, modular, lightweight (~65 KB), and instrument-agnostic solution that promotes integration and reproducibility within CRNP data analysis pipelines. In addition, CRNPy has a straightforward installation using the Python Package Index and minimal dependencies that are mostly included with the Anaconda open-source ecosystem. CRNPy's web documentation includes actual datasets and tutorials in the form of Jupyter notebooks that provide new users with an easily accessible entry point for CRNP data processing. The CRNPy library emphasizes easy maintenance and community-driven improvements since users can expand its capabilities by adding regular Python functions to the core module. The compact size and simple structure of the CRNPy library can also enable future integration into cloud-based services, IoT sensors, and system-on-chip technologies, broadening its use and customization potential.
 
 # Library features
 
