### 2024-04-22 soilwater/crnpy#10 @danpower101
@andres-patrignani there still needs to be some reference to previous packages in the paper, which at the moment is missing.

To keep the focus on the contributions of CRNPy, after describing the need for software solutions, a sentence could be added such as:

> "Python packages, such as `crspy` [@power2021cosmic] and `corny` [@cornish_pasdy], were previously developed in support of this. Here we present `CRNPy`, an alternative solution to support CRNP data processing. `CRNPy` was written with the aim to... [continue describing CRNPy]".

Perhaps @elbeejay could advise on this.
