### 2024-04-22 soilwater/crnpy#10 @andres-patrignani
@danpower101 I cannot agree more with you. I appreciate providing some tentative text to avoid the endless back and forth. I incorporated the suggested changes into the article. This new edit should conform with JOSS documentation and with the suggestions made by the editor and the reviewers. Please, let us know if there is anything else that we should modify. If not, I would appreciate closing this issue.
