### 2024-04-24 soilwater/crnpy#11 @elbeejay
Per https://github.com/openjournals/joss-reviews/issues/6025#issuecomment-2073812836 - @jlarsen-usgs has indicated that the suggestions made in this issue have been resolved.
