### 2024-04-24 soilwater/crnpy#10 @elbeejay
I think @danpower101 is spot on - it is important to acknowledge the other tools that exist in this space. JOSS does not require point-by-point or specific feature comparisons to be made, only that the need for the tool is clearly stated with respect to the other available software that already exists.

Thank you both for iterating on this.
