### 2024-04-24 #6025 @elbeejay
@joaquinperaza and @andres-patrignani - it looks like both reviewers have signed off on your revised manuscript at this time. I have a few minor editorial comments about the paper itself, line numbers refer to those in the latest PDF generated above.

- Line 8 has "cosmic-ray" and line 17 uses "Cosmic ray", I suggest making the inclusion or exclusion of the hyphen consistent
- Line 26: I suggest replacing "with" in the phrase "included with the Anaconda open-source ecosystem" for "within" or simply "in"

I will be generating a "post-review" checklist below that has final steps for you to confirm (or complete as necessary).

Please take this opportunity to re-read and make any final revisions to your paper in addition to the minor items I've highlighted above.

Thank you
