### 2024-04-25 709250c
Update paper.md
Made minor edits suggested by the editor. Adopted "Cosmic-ray neutrons" instead of "cosmic ray neutrons" across the entire paper.
