+++ paper.md
@@ -1,8 +1,8 @@
 ---
-title: "CRNPy: An Open-Source Python Library for Cosmic Ray Neutron Probe Data Processing"
+title: "CRNPy: An Open-Source Python Library for Cosmic-Ray Neutron Probe Data Processing"
 tags:
 - Python
-- Cosmic Ray Neutron Probes
+- Cosmic-Ray Neutron Probes
 - Soil Moisture
 - Data Processing
 date: "01 June 2023"
@@ -35,11 +35,11 @@ editor_options:
 
 # Summary
 
-CRNPy is a Python library that facilitates the processing, analysis, and conversion of raw cosmic-ray neutron counts obtained with stationary and roving cosmic-ray neutron probes (CRNP) into volumetric soil water content. The CRNPy library includes routines for atmospheric, biomass, and road corrections, along with one-dimensional and two-dimensional filtering. The library extends its utility by offering horizontal and vertical footprint determination, uncertainty estimation, depth extrapolation operators, and routines to assist users with field calibration. The design of the CRNPy library prioritizes reproducibility, ease of use, and compatibility across instruments, facilitating its adoption by manufacturers, end users, and researchers working to integrate non-invasive soil moisture sensing in agricultural and hydrological applications. 
+CRNPy is a Python library that facilitates the processing, analysis, and conversion of raw neutron counts obtained with stationary and roving cosmic-ray neutron probes (CRNP) into volumetric soil water content. The CRNPy library includes routines for atmospheric, biomass, and road corrections, along with one-dimensional and two-dimensional filtering. The library extends its utility by offering horizontal and vertical footprint determination, uncertainty estimation, depth extrapolation operators, and routines to assist users with field calibration. The design of the CRNPy library prioritizes reproducibility, ease of use, and compatibility across instruments, facilitating its adoption by instrument manufacturers, researchers, and end users aiming to integrate non-invasive soil moisture sensing in agricultural and hydrological applications. 
 
 # Statement of Need
 
-Cosmic ray neutron probes (CRNP) are non-invasive soil moisture sensors that fill the niche between point-level and satellite sensors. However, the conversion of raw CRNP data into soil moisture requires multiple corrections and filtering steps that are described across various peer-reviewed articles. To circumvent this limitation and enhance reproducibility, Python packages such as crspy [@power2021cosmic] and corny [@cornish_pasdy] were previously developed. In this study we present CRNPy, an alternative solution to support CRNP data processing that offers a simple, modular, lightweight (~65 KB), and instrument-agnostic solution that promotes integration and reproducibility within CRNP data analysis pipelines. The CRNPy library has a straightforward installation using the Python Package Index and minimal dependencies that are mostly included with the Anaconda open-source ecosystem. CRNPy's web documentation includes actual datasets and tutorials in the form of Jupyter notebooks that provide new users with an easily accessible entry point for CRNP data processing. The CRNPy library emphasizes easy maintenance and community-driven improvements since users can expand its capabilities by adding regular Python functions to the core module. The compact size and simple structure of the CRNPy library can also enable future integration into cloud-based services, IoT sensors, and system-on-chip technologies, broadening its use and customization potential.
[…]
