+++ setup.py
@@ -3,7 +3,7 @@ import setuptools
 
 setuptools.setup(
     name="crnpy",
-    version="0.6.0",
+    version="0.6.1",
     packages=['crnpy'],
     package_dir = {"": "src"},
     description="A Python package for the estimation and processing of soil moisture data from cosmic-ray neutron counts.",
