+++ src/crnpy.egg-info/SOURCES.txt
@@ -1,11 +1,0 @@
-LICENSE
-README.md
-setup.py
-src/crnpy/__init__.py
-src/crnpy/crnpy.py
-src/crnpy/data.py
-src/crnpy.egg-info/PKG-INFO
-src/crnpy.egg-info/SOURCES.txt
-src/crnpy.egg-info/dependency_links.txt
-src/crnpy.egg-info/requires.txt
-src/crnpy.egg-info/top_level.txt
