+++ src/crnpy.egg-info/requires.txt
@@ -1,5 +1,0 @@
-numpy
-pandas
-scipy
-requests
-utm
