### 2024-04-30 #6025 @joaquinperaza
Version 0.6.1 has also been archived in zenodo: [![](https://zenodo.org/badge/DOI/10.5281/zenodo.11090077.svg)](https://zenodo.org/records/11090077)

@andres-patrignani and I went over the text and checked the authors list, affiliations and ORCIDs. Please let me know if anything is still pending from our end. Thanks for your dedication and thanks @jlarsen-usgs and @danpower101 for reviewing our work.
