### 2024-05-06 #6025 @elbeejay
@joaquinperaza - I do have a few requests for you:

Please check for DOIs for the missing references as shown above, for example the "Re-examination of the First Five Ground-Level Events" citation has a DOI: http://dx.doi.org/10.22323/1.358.1149

Pandas should have a DOI as well: https://pandas.pydata.org/about/citing.html

As should matplotlib: https://matplotlib.org/stable/users/project/citing.html

I also need to ask that you revise the title of the Zenodo archive to match the JOSS paper title exactly. We also ask that the author names on Zenodo march what is in the JOSS paper (you can alter either the Zenodo archive or the JOSS paper to match middle names/initials).

Thanks, let me know when you get a chance to take care of these items.
