### 2024-05-07 f681e0e
Corrected authors and added missing DOI's
