+++ paper.bib
@@ -117,7 +117,8 @@ year = {2018}
   booktitle={International Cosmic Ray Conference (ICRC2019)},
   volume={36},
   pages={1149},
-  year={2019}
+  year={2019},
+  doi={10.22323/1.358.1149}
 }
 
 @article{dong2018soil,
@@ -291,8 +292,10 @@ DOI = {10.5194/hess-21-5009-2017}
   number={9},
   pages={1--9},
   year={2011},
-  publisher={Seattle}
+  publisher={Seattle},
+  doi={10.5281/zenodo.3509134}
 }
+
 
 @article{virtanen2020scipy,
   title={SciPy 1.0: fundamental algorithms for scientific computing in Python},
@@ -307,8 +310,16 @@ DOI = {10.5194/hess-21-5009-2017}
 }
 
 @article{hunter2007matplotlib,
-  title={The matplotlib user’s guide},
-  author={Hunter, John and Dale, Darren},
-  journal={Matplotlib 0.90. 0 user’s guide},
-  year={2007}
+  Author    = {Hunter, J. D.},
+  Title     = {Matplotlib: A 2D graphics environment},
+  Journal   = {Computing in Science \& Engineering},
+  Volume    = {9},
+  Number    = {3},
+  Pages     = {90--95},
+  abstract  = {Matplotlib is a 2D graphics package used for Python for
+  application development, interactive scripting, and publication-quality
+  image generation across user interfaces and operating systems.},
+  publisher = {IEEE COMPUTER SOC},
+  doi       = {10.1109/MCSE.2007.55},
+  year      = 2007
 }
