+++ paper.bib
@@ -284,18 +284,15 @@ DOI = {10.5194/hess-21-5009-2017}
   doi={10.1038/s41586-020-2649-2}
 }
 
-@article{mckinney2011pandas,
-  title={pandas: a foundational Python library for data analysis and statistics},
-  author={McKinney, Wes and others},
-  journal={Python for high performance and scientific computing},
-  volume={14},
-  number={9},
-  pages={1--9},
-  year={2011},
-  publisher={Seattle},
-  doi={10.5281/zenodo.3509134}
+@InProceedings{ mckinney2011pandas,
+  author    = { {W}es {M}c{K}inney },
+  title     = { {D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython },
+  booktitle = { {P}roceedings of the 9th {P}ython in {S}cience {C}onference },
+  pages     = { 56 - 61 },
+  year      = { 2010 },
+  editor    = { {S}t\'efan van der {W}alt and {J}arrod {M}illman },
+  doi       = { 10.25080/Majora-92bf1922-00a }
 }
-
 
 @article{virtanen2020scipy,
   title={SciPy 1.0: fundamental algorithms for scientific computing in Python},
