### 2024-05-07 #6025 @joaquinperaza
Thanks @elbeejay, I have addressed the corrections in the `JOSS` branch.
