### 2024-05-07 #6025 @elbeejay
@joaquinperaza the zenodo archive is missing the hyphen between "Cosmic" and "Ray" that is present in the `paper.md` file, please revise one or the other so that they are consistent. Thank you
