### 2024-05-08 #6025 @joaquinperaza
@elbeejay I fixed the archive title to be consistent with JOSS manuscript title
