### 2024-05-09 #6025 @andres-patrignani
@elbeejay I checked the final proof and everything looks good. We want to also thank you and the two reviewers (@jlarsen-usgs and @danpower101) for the constructive feedback and guidance, which greatly enhanced the quality of the library and the associated manuscript. The review process on Github was very dynamic, providing an excellent and refreshing reviewing experience.
