### 2024-05-15 2d6ed7a
Update paper.bib
Fixed capitalization of words
