+++ paper.bib
@@ -229,7 +229,7 @@ DOI = {10.5194/hess-21-5009-2017}
 }
 
 @article{hawdon2014calibration,
-  title={Calibration and correction procedures for cosmic-ray neutron soil moisture probes located across Australia},
+  title={Calibration and correction procedures for cosmic-ray neutron soil moisture probes located across {Australia}},
   author={Hawdon, Aaron and McJannet, David and Wallace, Jim},
   journal={Water Resources Research},
   volume={50},
@@ -295,7 +295,7 @@ DOI = {10.5194/hess-21-5009-2017}
 }
 
 @article{virtanen2020scipy,
-  title={SciPy 1.0: fundamental algorithms for scientific computing in Python},
+  title={SciPy 1.0: fundamental algorithms for scientific computing in {Python}},
   author={Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and others},
   journal={Nature methods},
   volume={17},
