### 2024-05-19 #6025 @joaquinperaza
@kthyng I have address the issue with my name in the Zenodo archive and @andres-patrignani have solved the issues in the JOSS paper.
