### 2024-05-24 #6025 @andres-patrignani
@kthyng I noticed that the citation in JOSS shows as "A. et al., (2024)", but it should be "Peraza Rud et al., (2024)".  If we fix/upload the .cff file, will this change? Any suggestions on how to get the citation correct would be greatly appreciated.
