### 2024-05-29 #6025 @xuanxu
Citation string fixed!
